Move NewGroupFragment to fragments.create
